+++ src/lib.rs
@@ -1,6 +1,11 @@
 pub mod common;
 pub mod component_store;
 pub mod components;
-pub mod gui_vizia;
 pub mod simulator;
+
+// Vizia frontend
+#[cfg(feature = "vizia")]
+pub mod gui_vizia;
+
+#[cfg(feature = "vizia")]
 pub use vizia;
